+++ crates/miden-protocol/src/transaction/verifier.rs
@@ -42,7 +42,7 @@ impl TransactionVerifier {
         );
         let stack_outputs = TransactionKernel::build_output_stack(
             transaction.account_update().final_state_commitment(),
-            transaction.account_update().account_delta_commitment(),
+            transaction.account_update().account_patch_commitment(),
             transaction.output_notes().commitment(),
             transaction.fee(),
             transaction.expiration_block_num(),